mise a jour

+++ controller/banquecontroller.js
@@ -1,5 +1,5 @@
 const Banque = require("../model/banque");
-const { updateChequePaiementStatus } = require("../controller/chequecontroller");
+const { updateChequePaiementStatus, updateChequeStatusToNon } = require("../controller/chequecontroller");
 const sendEmail = require('../model/mailer');
 const User = require('../model/user');
 const Offre = require('../model/offre');
@@ -7,16 +7,15 @@ const Cheque = require("../model/cheque");
 
 async function add(req, res, next) {
   try {
-    const { chequeId } = req.body;
+    const chequeId = req.params.chequeId;
 
     if (!chequeId) {
-      return res.status(400).send("L'ID du chèque est requis");
+      return res.status(400).json("L'ID du chèque est requis");
     }
 
-    // Vérifier si le chèque existe déjà dans la collection Banque
     const existingBanque = await Banque.findOne({ chequeId });
     if (existingBanque) {
-      return res.status(400).send("L'ID du chèque existe déjà dans la base de données");
+      return res.status(400).json("L'ID du chèque existe déjà dans la base de données");
     }
 
     const banque = new Banque({ chequeId });
@@ -26,22 +25,22 @@ async function add(req, res, next) {
 
     const cheque = await Cheque.findById(chequeId).populate('factureId');
     if (!cheque) {
-      return res.status(404).send("Chèque non trouvé");
+      return res.status(404).json("Chèque non trouvé");
     }
 
     const facture = cheque.factureId;
     if (!facture) {
-      return res.status(404).send("Facture non trouvée");
+      return res.status(404).json("Facture non trouvée");
     }
 
     const offre = await Offre.findById(facture.offreId);
     if (!offre) {
-      return res.status(404).send("Offre non trouvée");
+      return res.status(404).json("Offre non trouvée");
     }
 
     const user = await User.findById(offre.userId);
     if (!user) {
-      return res.status(404).send("Utilisateur non trouvé");
+      return res.status(404).json("Utilisateur non trouvé");
     }
 
     await sendEmail(
@@ -50,10 +49,10 @@ async function add(req, res, next) {
       `Bonjour ${user.firstName} ${user.lastName},\n\nVotre chèque pour l'offre ${offre.nom} a été encaissé dans la banque.\n\nMerci.`
     );
 
-    res.status(201).send("Banque ajoutée avec succès et email envoyé");
+    res.status(201).json("Banque ajoutée avec succès et email envoyé");
   } catch (err) {
     console.log(err);
-    res.status(500).send("Erreur lors de l'ajout de la banque");
+    res.status(500).json("Erreur lors de l'ajout de la banque");
   }
 }
 
@@ -73,13 +72,23 @@ async function update(req, res, next) {
     res.send("updated");
   } catch (err) {
     console.log(err);
-    res.status(500).send("Erreur lors de la mise à jour de la banque");
+    res.status(500).json("Erreur lors de la mise à jour de la banque");
   }
 }
 
 async function deletebanque(req, res, next) {
   try {
+    const banque = await Banque.findById(req.params.id);
+    if (!banque) {
+      return res.status(404).send("Banque non trouvée");
+    }
+
+    const chequeId = banque.chequeId;
+
     await Banque.findByIdAndDelete(req.params.id);
+
+    await updateChequeStatusToNon(chequeId);
+
     res.send("deleted");
   } catch (err) {
     console.log(err);
