+++ controller/chequecontroller.js
@@ -5,7 +5,8 @@ const moment = require('moment');
 
 async function addCheque(req, res, next) {
   try {
-    const { factureId, montant } = req.body;
+    const { montant } = req.body;
+    const factureId = req.params.factureId;
 
     const facture = await Facture.findById(factureId);
     if (!facture) {
@@ -21,7 +22,7 @@ async function addCheque(req, res, next) {
       return res.status(400).send("Le montant du chèque dépasse le reste de la facture");
     }
 
-    const cheque = new Cheque(req.body);
+    const cheque = new Cheque({ ...req.body, factureId });
     cheque.paiement = "non";
     await cheque.save();
 
@@ -33,7 +34,7 @@ async function addCheque(req, res, next) {
 
     await facture.save();
 
-    res.send("Chèque ajouté");
+    res.json("Chèque ajouté");
   } catch (err) {
     console.log(err);
     res.status(500).send("Erreur lors de l'ajout du chèque");
@@ -53,9 +54,7 @@ async function show(req, res, next) {
 async function update(req, res, next) {
   try {
     await Cheque.findByIdAndUpdate(req.params.id, req.body);
-
     await updateChequePaiementStatus(req.params.id);
-
     res.send("updated");
   } catch (err) {
     console.log(err);
@@ -65,8 +64,30 @@ async function update(req, res, next) {
 
 async function deletecheque(req, res, next) {
   try {
+    const cheque = await Cheque.findById(req.params.id);
+    if (!cheque) {
+      return res.status(404).send("Chèque non trouvé");
+    }
+
+    const factureId = cheque.factureId;
+    const montant = cheque.montant;
+
     await Cheque.findByIdAndDelete(req.params.id);
-    res.send("deleted");
+
+    const facture = await Facture.findById(factureId);
+    if (facture) {
+      facture.montantCheque = (facture.montantCheque || 0) - montant;
+
+      if (facture.montantApresRemise - facture.montantCheque !== 0) {
+        facture.statut = "non soldé";
+      }
+
+      await facture.save();
+    }
+
+    await updateChequePaiementStatus(req.params.id);
+
+    res.json("Chèque supprimé");
   } catch (err) {
     console.log(err);
     res.status(500).send("Erreur lors de la suppression du chèque");
@@ -77,19 +98,23 @@ async function checkChequeEcheance(req, res, next) {
   try {
     const cheques = await Cheque.find();
     const today = moment();
+    let countChequesDue = 0;
 
-    const chequesWithStatus = cheques.map(cheque => {
+    cheques.forEach(cheque => {
       const echeanceDate = moment(cheque.echeance);
       if (today.isAfter(echeanceDate)) {
-        cheque.statut = 'échu';
-      } else {
-        const joursRestants = echeanceDate.diff(today, 'days');
-        cheque.statut = `Il reste ${joursRestants} jours`;
+        countChequesDue++;
       }
-      return cheque;
     });
 
-    res.json(chequesWithStatus);
+    const totalCheques = cheques.length;
+    const percentageDue = (countChequesDue / totalCheques) * 100;
+
+    res.json({
+      totalCheques: totalCheques,
+      chequesDue: countChequesDue,
+      percentageDue: percentageDue.toFixed(2) + "%"
+    });
   } catch (err) {
     console.log(err);
     res.status(500).send("Erreur lors de la vérification des échéances des chèques");
@@ -106,8 +131,25 @@ async function updateChequePaiementStatus(chequeId) {
     }
   } catch (err) {
     console.log(err);
-    // Gérer l'erreur selon les besoins
   }
 }
 
-module.exports = { addCheque, show, update, deletecheque, checkChequeEcheance, updateChequePaiementStatus };
+async function updateChequeStatusToNon(chequeId) {
+  try {
+    await Cheque.findByIdAndUpdate(chequeId, { paiement: "non" });
+  } catch (err) {
+    console.log(err);
+  }
+}
+
+async function showNonPaidCheques(req, res, next) {
+  try {
+    const nonPaidCheques = await Cheque.find({ paiement: "non" });
+    res.json(nonPaidCheques);
+  } catch (err) {
+    console.log(err);
+    res.status(500).send("Erreur lors de la récupération des chèques impayés");
+  }
+}
+
+module.exports = { addCheque, show, update, deletecheque, checkChequeEcheance, updateChequePaiementStatus, updateChequeStatusToNon, showNonPaidCheques };
